pushitfreelance/requirements.txt
django
djoser
djangorestframework
markdown
django-filter
drf-yasg
drf-nested-routers